Extension gets authenticated user from service
The Perforce extension now sends the user to the authentication service on
login and then pings the service to find out if the user has successfully
authenticated with the OIDC provider. It then compares the email address to
ensure it matches, which is not much of an assertion, but for now it
suffices to demonstrate the procedure works.

Still need to secure the channel between the extension and service.