Add extension settings for user/profile mapping
To facilitate a flexible approach to matching Perforce users to the profile
returned from the identity provider, the extension now has two settings that
are configurable per installation.

user-identifier: the name of the Perforce trigger variable, "user" or "email"
name-identifier: the name of the profile field to use in matching

Both default to "email" which is likely to work well with OpenID Connect. A
likely choice for SAML would be "user" and "nameID", if you had Perforce user
names that matched the identifiers used in the IdP.