SAML legacy support for Swarm 2018.3
The auth service can act as a SAML IdP such that Swarm will use the auth
service as the SAML IdP, and ultimately authenticate with Perforce. There is
some configuration on the service side, and a small hack in which the
service assumes that the nameID is also the "user identifier" on the login
extensions side. This hack may be eliminated if Swarm sends the SAML
response as the token instead of the user name, which it seems to be doing
in testing.

The auth service assumes that the client initiating a SAML login wants the
user to authenticate with SAML. For OpenID Connect, some additional work
will be necessary, namely a "default" protocol setting, and a means for
redirecting to OIDC for login, and back again to the SAML IdP code.