Detect unset extension configuration, validate
Add a leading ellipsis to extension configuration values, which serve as
documentation, to make it easy for the extension to determine if the
settings have been explicitly set. This is especially difficult for the list
of users and groups, since they are whitespace-separated words and may have
uppercase and/or punctuation (technically speaking). With the illegal
wildcard prefix, it will be easy to ignore undefined configuration.

Also perform some basic validation of the global settings.