Fix logout when using Swarm
When logging out from Swarm, we can direct it to the auth service, and then
the auth service can properly log the user out of the identity provider.
This is fine, except if we are using OIDC then the user profile most likely
does not have a field named 'nameIDFormat', which the SAML library requires.
As such, we now set that field to the 'unspecified' constant if it is
missing. This also helps of the SAML IdP were to fail to set the field,
unlikely as that may be.