Add SAML_NAMEID_FIELD envar to set SAML IdP nameID
For the SAML IdP behavior, it may be desirable to specify the field in the
user profile to use if 'nameID' is not already present (e.g. if using OIDC
with the actual IdP). That is, if 'nameID' is not available, this
environment variable provides the name of the property to be selected. If
not specified, the application will try "preferred_username" and "email",
which are supplied by Okta. Auth0 has "nickname" and "name" as well. Not
sure about OneLogin, my account seems unusable at this point.