Make some SAML IdP adjustments for HTH
Note that TeamHub does not use the SAML response to authenticate with Helix
Server, so that integration is not yet available. For now at least, the
service can process the SAML request from HTH and provide an acceptable
response. In the future, HTH may allow for authentication with Helix using a
SAML response, as Swarm does now.