Delete user profile upon retrieval
Prevents a replay attack on a user that has recently authenticated with the
service. For several minutes after a successful login, the user profile data
is cached. Another user could attempt to log in as that user and would be
successful, despite not providing any credentials to IdP. With this change,
once the extension requests the user profile data, it is removed from the
cache, preventing a replay attack.

Jobs: AS-3