Bootstrap OIDC strategy on demand
In an effort to simplify the docker setup, and remove the depends_on from
the service container, the service now loads the OIDC strategy when it is
first needed, rather than on startup. This also prevents writing an error to
the log on startup because the OIDC IdP is not configured (which would be
fine if the service is not using OIDC).