Allow forcing authentication in OIDC per request
As with SAML, the client can force authentication in OIDC by setting the
forceAuth query parameter when starting a login request, or by setting
FORCE_AUTHN environment variable to any non-empty value. This works with
Auth0 but apparently is not supported in Okta or OneLogin when using OIDC,
nor does it work with the oidc-provider docker container.