Default the SAML NameID format to unspecified
The NameID format of "unspecified" works with all of the tested providers,
including Auth0, Okta, OneLogin, and PingFederate. The original default,
defined by the passport-saml library, was "emailAddress", which worked for
most providers, but not the default configuration when using PingFederate.

In any case, the value can be configured easily using environment variables.