Infer port, protocol from SVC_BASE_URI by default
If PORT or PROTOCOL are not set in the environment, then use the result of
parsing the SVC_BASE_URI value instead. This allows the old settings to
continue to take effect, while also eliminating the need to set them at all.

That is, do not set PORT or PROTOCOL in the environment anymore. The service
will infer both values from SVC_BASE_URI. If the port is not found in the
URI, and the protocol is http: then the port will default to 80, and if the
protocol is https: then it will default to 443.

Remove the now unnecessary setting of PROTOCOL in several places.