Use the SP certs instead of redundant "IdP" certs
Without thinking, I added another set of certificate definitions for the
case where the authentication service would act as an IdP to clients such as
Swarm. However, this is completely unnecessary, and we can use the existing
"SP" certificates for this purpose. There is nothing special about them,
they simply act as certificates representing the auth service -- we only
need one set of such certificates.