Make auth service URL configuration orthogonal
Auth service will return 'loginUrl' and 'baseUrl' in the initial new request
response. The client making the new request can use these values to produce
a URL for the end user to log in. As a result, the server extension no
longer needs the administrator to set the Auth-Protocol value. Likewise, the
Service-URL extension setting can be an internal address by which the
extension will connect to the service. The extension will only use that URL
to connect to the service, and return the 'loginUrl' provided by the service
to the user.

If the administrator sets the Auth-Protocol setting in the extension, then
the extension will add that to the 'baseUrl' provided by the service. This
is helpful because it means that the administrator only needs to get the
user-visible auth service URL setting correct in one location, the auth
service configuration (and the IdP, but we have no control over that).

This change is backward compatible. Existing clients can keep on doing what
they were doing, but this is a better strategy going forward.