Handle reload of certain pages better
As mentioned on the Blitz page, reloading certain pages would result in an
error message displayed on the page in the browser. For instance, log in
using SAML, then reload the "success" page -- it would show an error.
Similarly, entering /saml/logout two times in a row would result in an
error.