Fix Shibboleth config to return email as nameID
While the initial setup was "working" for the authentication service, it was
not set up to return an email address as the nameID. In fact, it returned
nothing for the nameID, so the Helix Server extension had nothing at all by
which to authenticate the user. This configuration now fixes that.