Prohibit install by root, check file perms
The service install.sh should not be run as the root user, otherwise when
Node attempts to install the dependencies, misleading file permissions
errors can occur. The script itself will use sudo as needed, so running as
the root user is not necessary.

Additionally, check that the user running the script can write to the
service directory by preemptively creating the node_modules directory. If
that fails, display a clear and concise error message.