Support file globbing for finding CA files
Previously, the CA_CERT_FILE setting specified a single file, if any at all,
and the service would read this file and pass that as the trusted
certificate authority to Node.js (used in verifying the clients requesting
user data, e.g. the Helix Server extension).

With this change, CA_CERT_FILE can now be a file glob, meaning it has
wildcards that match multiple files. The service will then read all of the
files matching the pattern and pass them all to Node.

See https://github.com/isaacs/node-glob for the details of the support
patterns, which include the obvious '*' for matching multiple characters,
and '?' for matching a single character.

While previously you might set CA_CERT_FILE to 'certs/ca.crt' to read that
one file, you can now use 'certs/ca*.crt' to find all files in the certs
directory that starts with 'ca' and ends with '.crt', as a simple example.

This change should not have any effect if the CA_CERT_FILE value does not
have wildcard characters (*, ?, [], !(), and so on) which is likely the case
because most of these are not valid file name characters anyway.