Permit specifying the SAML IdP certificate
This provides an additional layer of security by permitting the HAS
administrator to save a local copy of the SAML IdP certificate and pass that
to HAS. Thus HAS can validate the IdP is the one trusted by the
administrator, rather than blindly trusting the metadata returned from the
IdP at the time of authentication.

The SAML IdP certificate file is specified with the IDP_CERT_FILE
environment variable, and must be a PEM-encoded X.509 signing certificate.
If not provided, the previous behavior of relying on the IdP metadata will
take effect.