Allow setting authn context other than the default
In some cases, it seems that the IdP does not tolerate the default authn
context value of PasswordProtectedTransport when using SAML authentication.
One instance involved Azure AD, in which the configuration of the password
database (?) prevented the user from authenticating using this default
context. This change allows the HAS administrator to change the context to
one that will work, such as urn:oasis:names:tc:SAML:2.0:ac:classes:Password
or WindowsIntegrated (?).

Fixes AS-29