doc: give guidance on configuring OIDC issuer URI
It was not clear how to find the OIDC issuer URI and at least one customer
had trouble and attempted to use an "endpoint" URL instead. The section on
configuring OIDC now clearly states how to find the correct URI.

Also rearranged some of the content, renamed some headings, and changed the
heading level across the board to be more consistent.