Allow logging to console, file, or syslog
This change introduces a new environment variable named LOGGING, which is
documented in the administrator's guide. With this setting the administrator
can now direct logging messages to the console, or to a named file, or to
the system logger (i.e. syslog).

The DEBUG environment variable still works as before: setting it to any
value will enable full debug logging to the console. If neither DEBUG nor
LOGGING are set, only HTTP requests are written to the console, as before.