Allow SAML to configure via IdP metadata
By setting the SAML_IDP_METADATA_URL environment variable, the service will
configure some of the SAML settings via the identity provider's metadata.
This is purely optional as not all customers will want to take the risk of
trusting data provided over the network. However, it is a little bit easier
to configure SAML using the IdP metadata in many cases.