Fix SAML login after subtle passport-saml change
The 1.3.0 release of passport-saml changed the handling of the private
certificate, and as a result, it broke all SAML logins. The fix is to read
both the private cert and the decryption key using UTF-8 encoding such that
the result is a String object instead of a Buffer.

Tested with Azure, Auth0, Okta, OneLogin, and Shibboleth.