Enable filtering client certs by common name
Per customer request, the /requests/status REST API now can accept or reject
client requests based on the common name given in the client certificate.
This is controlled by the presence of the CLIENT_CERT_CN environment
variable. If present, it is used as a simple regular expression for matching
the common name of the client certificate.

The exact details of the pattern matching are described at the minimatch
project page: https://github.com/isaacs/minimatch -- however, there is no
need to document _all_ of those options. It is likely more than sufficient
to provide the common name as-is (e.g. "TrustedClient") or with a single
wildcard (e.g. "*.trusted.com"). For customers with more sophisticated
needs, they can contact support to learn the many available options.

npm test passes