Use Clean Architecture design to improve testing
See the docs/Design.md for details and explanation. In short, these changes
allow for more thorough testing of the business logic of the service. What's
more, the policy and details of the application have been separated, making
it easier for others to comprehend the code and discern the data flow.

There should be no user-visible changes, and none of the existing tests were
changed to accomodate the new design.

npm test passes