Fix OIDC authorization code flow support
In short this means using PKCE and the authorization code flow rather than
the implicit code flow, which is less secure and now considered deprecated.

Tested with Auth0, Azure, Okta, and OneLogin.

Not sure if it is related, or a change in Azure, but the OIDC response from
AAD no longer includes the user email address. Tried various settings in
Azure to no avail. Fortunately, can configure the extension to use "name"
and "fullname" to verify the user authentication.

npm test passes