Fold user profile into SAML request validation
When using Swarm with HAS via SAML, the p4d extension receives the SAML
response as the user's password. This is sent back to HAS for verification,
and that yields the available information contained in the SAML response.
However, when using an OIDC provider with HAS, the information that would
come back to the extension via the normal authentication route (i.e. when
authenticating with p4 or p4v) typically contains other information. With
SAML the profile would very likely have nameID but with OIDC there is
usually an email, an no nameID property. As such, the extension cannot work
with both OIDC and SAML responses coming back from HAS.

This change addresses that by merging the user profile data associated with
the nameID into the profile that is returned from the validate endpoint.

npm test passes