Ensure logout takes the correct route for SAML
When the SAML protocol handler is acting as the IdP (to Swarm), it is quite
possible that the default protocol for the service is OIDC. In this
situation, the /saml/logout route should be redirected to the other protocol
handler (OIDC) to ensure correct behavior. This was being done already, but
it was done in a way that was fragile. If the DEFAULT_PROTOCOL was set to
"oidc" but the client was logging in using SAML anyway (i.e. as with the
automated tests, and quite possibly the client application), then this
behavior would in fact be wrong.

This change establishes the logout redirect only if the login was initiated
by a SAML login request (i.e. Swarm), and the default protocol is defined,
and it is not "saml". Otherwise, the normal logout behavior will be used.

npm test passes