Fix IdP metadata cert to work with passport-saml
When using a metadata URL to configure a SAML identity provider within the
authentication service, if the metadata includes a `cert` property that is
an array of strings, or is a string that is PEM formatted, the passport-saml
library will fail to authenticate with the identity provider.

This can be fixed by converting an array to a single string, and by changing
the string formatting to a single line without a PEM header and footer.

This change happens to allow the Shibboleth SAML IdP in docker container to
be configured within the auth service using just the IdP metadata URL. And
that will help with writing additional automated tests for SAML.

Tested manually with Auth0, Azure, Okta, and OneLogin.

npm test passes