Move recently added admin material to new guide
All new features and settings added after the 2019.1 release should be in a
separate version of the admin guide. The existing guide, which has v2019.1
in the name, should only reflect features and settings available in that
release. The troubleshooting and other helpful elements will remain.

The new guide is temporary and eventually the admin guide will exist only on
the Perforce web site rather than this Markdown file.