Ensure pm2 startup success by specifying systemd
It had been working previously, but now it seems that pm2 must have the
startup system specified on the command line, otherwise it gets confused
about the arguments and fails during the package post-install. Since the
default for Ubuntu 16 and 18 and CentOS 7 and 8 is "systemd", will use that
as the desired startup manager.