Prefer "entity ID" over "issuer" in SAML settings
It seems to be the case that the term "issuer" may not be as common as
"entity ID", so the SAML settings have been changed as follows:

* The new SAML_SP_ENTITY_ID setting will work the same as the old SAML_SP_ISSUER.
* The new SAML_IDP_ENTITY_ID setting will work the same as the old SAML_IDP_ISSUER.

Also it seems that the default value of urn:example:sp appears to be magical
and it is unclear if the value should be modified. The suggestion was made
that the default should be a URL, which is often preferred anyway. As such,
the default is now https://has.example.com instead of urn:example:sp in the
example configuration file.

The existing Markdown admin guide has been updated.

npm test passes