Add configure script to configure auth service
Like the configure script for Helix Core, Swarm, and others, this shell
script is suited for Linux systems, is both interactive and scriptable,
and will configure the auth service and restart as needed.

No automated testing as yet. Once that, and QA vetting has occurred,
references to this script from the install script will be added.