fix: validate OIDC inputs more accurately
The OIDC issuer URI has to be an HTTPS URL according to the spec. Previously
any value was accepted, which is not at all correct.

If the user inputs an OIDC issuer URI, then the client key and secret are
required, where previously they were always optional during prompting but
then resulting in a validation error later in the script.