fix: log subject of cert, if known, for debugging
Log information about the client certificate sent, or not sent, in the
request for the user login status. If no certificate is available, clearly
state that. If there is a cert, show the common name of the subject, or
'(none)' if the subject is not provided. This information can help with
diagnosing authentication failures in which the service responds with a 401
but the client is sure to be sending a client certificate.

npm test passes