config: improve config script based on feedback
Based on feedback from members of the consulting team, the configure script
now offers a bit more explantory text for some inputs, as well as examples
where appropriate. Additionally, the interactive mode will prompt for which
protocol to configure, either OIDC, SAML, or both. The configure script now
configures the logging to emit debug level messages to files that are
limited in size and number with automatic rotation.