configure: do not prompt for provided secret twice
If the user provided the OIDC client secret using command line arguments,
and is running the script in interactive mode, the prompts will display the
given values. The user then simply presses Enter to accept the already
provided values. However, with the client secret, the user was getting
prompted twice, which is of no use since they did not type in the password
at the prompt, so there is nothing to verify.