fix: recommend pm2 kill/start over startOrReload
When the service is running in "production" mode, pm2 will not reload the
configuration when invoked with the startOrReload command. Since the
configuration is set to production by default, the recommended command in
the install.sh and admin guide should be corrected. Now both the install
script and the admin guide recommend using pm2 kill and pm2 start.