fix: detect startup system for pm2
The package post-install scripts were assuming that the startup system is
always going to be systemd. This may be true most of the time, but it is not
guaranteed since the admin may change to something else (initd). This change
to the post-install scripts is such that it invokes pm2 once to detect the
actual startup system, and then use that instead of assuming systemd.