fix: add python3 to avoid "error:" in build output
The automated build system detects the "error:" in the output of the test
process and flags the build as a failure. To avoid this, install the python3
package before running the install/configure scripts. The fact that python
is missing does not change the result, but the output suggests that
something went wrong, which is not really true.