fix: revert certificates change to support CentOS7
Orignally the self-signed certificates were generated using openssl and the
default SHA1 message digest. This caused a failure when the extension
connected to the auth service when running on Debian buster. To fix that,
the certs were generated using the SHA256 message digest. That however,
causes a (silent) failure when running on CentOS 7.8, one of our supported
platforms. As such, this change reverts the certificates to those signed
with the SHA1 message digest. A note will be added to the loginhook
extension admin guide to address this issue.

npm test passes