fix: set session cookie expires to one day
The session cookie for HAS was using mostly default settings, and among
these was the expires/max-age value, which was left unset. This means the
session cookie had no expires at all, and this seemed fine for a long time.
However, in some cases, or with recent versions, Chrome tosses the cookie
away in the middle of the login process. When a successful SAML response
comes back to HAS from the IdP, the session cookie is already gone,
resulting in the error "request identifier must be defined".

This change sets the cookie to expire in one day. It also sets the "rolling"
property to true, which means the expires date will be bumped on each
request, so the cookie will not expire if the browser makes a request each
day (not that this is necessary, but it means this issue will be much less
likely to appear again).

npm test passes