fix: turn off colors in logging output
The color codes used to make the log output colorful on a console, typically
end up in log files because the output is captured by the process manager.
Then when viewing the log snippets, the output is more difficult to read
because of the cryptic color codes (e.g. ^[[31m) on every line. Since they
are of no functional value and only make the log output more difficult to
read, this change turns off all color output in the logs.

npm test passes