fix: avoid exception by pruning empty cert entries
When setting either CA_CERT_FILE or CA_CERT_PATH in the configuration, it is
possible for the service to process an empty directory. When this happens,
an exception would be raised (and the service likely did not function
correctly). With this change, any empty results found during processing of
those files and directories will be removed before proceeding.

npm test passes