build: introduce use of eslint to catch JS issues
ESLint (https://eslint.org/) is a JavaScript static analysis tool which
finds many coding problems that could become issues at run time. For now,
the tool is installed as a dev dependency and can be run like so:

$ npx eslint lib routes

Added several line-disables for unused variables, which is unavoidable for
callbacks whose definition is outside of our control.

npm test passes