fix: SAML detection was broken for help message
On the index page, if the service was not set up with either OIDC or SAML,
it would should a message about configuring the service. This used to work
fine, but now SAML can be configured with a different environment variable,
and the message was appearing by mistake. Now that logic has been fixed.

npm test passes