fix: change the session settings for Chrome
Seems that Google Chrome treats the auth service's session cookie as a
"third party" cookie because it is not marked secure when on the HTTPS
protocol. To fix this, make sure the cookie is marked as secure when using
HTTPS for the service.

Also make the session cookie name generic to avoid advertising the identity
of the service to attackers.

npm test passes