test: ubuntu docker images lost USER envar
Published releases of Ubuntu docker images are not stable, as it turns out,
and recently the released versions of the docker images changed such that
they lost the USER environment variable. Wow.

Also tweak the regex for matching the version in the package.json file so
that it matches the development versions defined in the test docker file.