Add example container using HAProxy
This is merely an example container, and is not configured by default when
setting up HAS and the other docker containers. By applying the code patch
and starting up the containers, you will have HAProxy in front of HAS, and
it will terminate the SSL connection and mark the session cookie as secure,
even though HAS is not running on HTTPS. The drawback is that HAS is not
running on HTTPS and therefore not verifying the client requests for the
user profile data.