feat: offer redis support for session state
In addition to the existing in-memory data store for both the browser
session state and the request/user mapping, the application now offers
support for using an external Redis instance. This support is enabled by
setting the REDIS_URL environment variable, which points to the Redis
instance to be used for storing all of the data. This allows for easy
failover of the application, via a load balancer, assuming each instance is
configured to use the same Redis instance.

npm test passes