fix: use dash as separate in package version
The Linux package build was using a dot to separate the "release base" and
the "patch" values when updating the "version" property in the package.json
file. This could result in a version like "2020.1.1.9999999" which is not a
valid Node.js package version, and would result in errors when invoking the
npm command. Note that this does not stop pm2 from running the service, but
it would prevent an admin from running npm commands on the service.

Instead of a dot, the build process now uses a dash as the separator, since
that is okay for Node.js package versions (e.g. "2020.1.1-999999" is okay).