fix: write to the console when server start fails
If the server is configured to bind to a port that the user running the
server does not have permission to (e.g. anything below 1024 can only be
bound to by a privileged user), and the logging is configured to write to a
file, then the startup fails miserably due to "errno 1" with no explanation
whatsoever. This is not at all helpful since "errno 1" merely means there
was a permissions problem. A file permission? Which file? No, it was a
socket bind error.

Now the error handling code for the socket bind writes to the console,
rather than whatever the logger is configured to do. This ensures that the
message is displayed somewhere, either in the console, or in the pm2 logs,
rather than disappearing into a black hole.