Allow rule-based routing w/o redis enable failover
While Redis is very helpful for supporting failover in the application, by
storing the session state and other data external to the HAS instances, it
does incur an additional cost in terms of deployment and configuration. For
those cases where seamless failover is not necessary, the auth service can
be configured to include a query parameter in the initial login URL. This
will allow the proxy or load balancer to employ a rule for routing the
request to the correct service instance. This is in addition to the cookie
based session affinity that should be used when Redis is not in use.

npm test passes