Add setup for publishing HAS to Docker Hub
The basic steps are shown in the containers/hub/Dockerfile file, with the
presumption that you have installed Docker Desktop and run `docker login`
already. The contents of the README.md file must be copied manually onto the
Docker Hub listing, since we do not have an automated build setup for this.