fix: start pm2 as the normal user rather than root
For consistency with the configure script, pm2 will be started as the user
that is invoking the install script. Since the script prohibits the root
user from running it, then the unprivileged user will be the owner of the
pm2 process. This is desirable for several reasons, and is now consistent
with the packages and configure script.