fix: remove trailing slash from SVC_BASE_URI
The unmodified value for SVC_BASE_URI was being returned in the new request
response, and the extension then tacked on its own path to the URL. If that
URL was provided by the user with a trailing slash, then the resulting URL
would have a double-slash. This results in a 404 since the extra slash
causes the router to fail to match in the service.

npm test passes