fix: upgrade Node.js to v14 is permitted by user
During the installation, if an older version of Node.js is discovered, and
it is from a 'nodesource' package, then offer the user an opportunity to
upgrade the package to the required v14 release.

If the install script is run in non-interactive mode, then a new option must
be provided (--upgrade) to perform the upgrade automatically.

In all other cases, the script exits with an error.