Replace module-alias with "The Wrapper" solution
Node.js is lacking when it comes to finding and loading "modules" such that
any deeply nested directory structure quickly becomes unwieldy. The previous
attempt to alleviate this pain was to use the module-alias module, and that
worked well enough. However, in attempting to compile this application into
a binary using the pkg tool, it was discovered that this custom module
loader causes problems with the resulting binary (errors and nothing loads
properly). For now, replace module-alias with a simple wrapper function that
builds the path to the specified file.

npm test passes