pkg: explicitly list all scripts and assets
The pkg tool cannot find files to include in the final binary if the require
call involves dynamic values. List the files and directories where we have
scripts and assets that need to be included. A couple of directories in the
saml and samlp module dependencies were also flagged during runtime.