pkg: configure script modifies .env or ecosystem
If the configure script can find the pm2 executable, then it assumes that it
should modify the ecosystem.config.js file, just as before. However, if pm2
is not easily discoverable, then the configure script will modify the .env
file instead, since the ecosystem file is only used by pm2.

Not having pm2 and not using the ecosystem.config.js file will be the norm
when using the packaged version of HAS based on the single binary build.