fix: log client cert errors to aid in debugging
The service was returning the client certificate validation error message to
the client but not logging it. Since the client is typically the extension
running in the server, and it is also not logging the message, no one ever
knew why the request was failing, exactly. Granted, the error is a 408 which
means the client cert was rejected, but the details were obscured.

npm test passes