Fix packages to tolerate missing systemd
The automated package tests rely on Docker, which has a hard time running
systemd since that requires privileged access. Given there is a chance that
the target system might not have systemd, the packages should allow for such
a scenario. As such, the Linux package post-install and pre-remove scripts
will check if systemd is functional before applying any service changes.