config: read .env settings to prime input values
The configure script will now read the .env file to gather any existing
settings, saving the admin the effort of entering the same values again on
subsequent invocations of the script. That is, after running the script
once, there will be a .env file with the saved settings. Running the script
again will show the existing settings as the default values for the various
input prompts. Pressing Enter will accept those defaults.