feat: reload most settings when USR2 signal received
While the service is running, if it receives the USR2 signal, then it will
reload most of the environment variables. Those settings that _cannot_ be
changed during runtime are: BIND_ADDRESS, CA_CERT_FILE, CA_CERT_PATH, DEBUG,
LOGGING, PROTOCOL, PORT, REDIS_URL, SESSION_SECRET, SVC_BASE_URI, and
TRUST_PROXY.

To send the USR2 signal to the process, get the process identifier of the
service and then invoke the kill command like so:

% kill -USR2 <pid>

This command and the correct process ID are written to the log at info level
when the service starts up.

npm test passes