fix: handle changing protocols on subsequent runs
When a configuration file exists and the configure script is run, it will
get stuck in an input validation loop if the user chooses to configure a
different protocol than the one defined previously. That is, the user
configured SAML the first time, saved the changes, then ran the script
again, but this time choose OIDC: the script gets stuck in a loop trying to
validate the inputs.