fix: add --pm2 option to configure, do not guess
The configure script no longer attempts to guess if it should modify the
ecosystem.config.js file or the .env file, and instead assumes that we
should be modifying the .env file. With the new --pm2 option, the script
will instead modify the ecosystem.config.js file, as before. This is part of
the move away from pm2 and its configuration file, and toward systemd and
the use of the .env file for configuring HAS.