fix: logging config path was wrong in many cases
The configure script was always setting the LOGGING value to a relative path
that started with '..', which assumed that HAS was always run as a Node.js
application via the bin/www script. However, that is not always the case, as
with the single-binary build installed via the package. The configure script
now considers whether the bin/www script exists or not when setting the
LOGGING value.