feat: support PKCS#12 and a private key passphrase
As an alternative to specifying a public/private key pair in PEM format
using the SP_CERT_FILE and SP_KEY_FILE settings, the administrator can
instead specify a PKCS#12 formatted public/private key pair in a single
file, with the new PFX_FILE setting.

Additionally, if the private key (regardless of file format) is encrypted
with a passphrase, the KEY_PASSPHRASE or KEY_PASSPHRASE_FILE settings can be
used to provide that passphrase to decrypt the private key at startup.

npm test passes