fix: process SAML_AUTHN_CONTEXT as encoded list
If the SAML_AUTHN_CONTEXT environment variable is defined, and it is a
string, and it contains square brackets (technically just '[' is enough),
then convert the string to a list of strings. That is:

SAML_AUTHN_CONTEXT='["...:X509", "...:MobileTwoFactorContract"]'
(a string containing an encoded list of strings)

will become:

SAML_AUTHN_CONTEXT=['...:X509', '...:MobileTwoFactorContract']
(a list of strings)

This is safe since a string with a single SAML authn context can never have
square brackets in it, so if this strings does contain them, than it must be
a list encoded as a string.

Note that the passport-saml library can take either a single string or a
list of strings for the authnContext parameter.

npm test passes