fix: require certs for SAML validation only for https
When validating the SAML response received by Swarm (and sent via the
extension), the service should only require client certificates when using
the HTTPS protocol. Previously it was always checking for certs, which is
never going to work when processing HTTP requests.

Additionally, the SAML response validation code was broken due to the recent
upgrade of the passport-saml library. The internal workings of that library
changed and as a result the SAML validation was failing and resulting in an
error.

npm test passes