test: sign service cert using fake authority
Rather than having a plainly self-signed certificate, use the self-signed
authority to sign the request for the service. This is in addition to the
extension, whose cert is already signed in this manner. Now the extension
can use the fake CA to verify the service certificate, by changing the two
`false` values to `true` at the end of curlSecureOptions() in the main.lua
file. For testing and development the verification will remain disabled.

npm test passes