fix: use latest node-saml library, fix API usage
The version of the node-saml library used previously was outdated with
respect to the related package, passport-saml, despite the version numbers
being the same. In the course of updating the dependency and fixing the API
usage, discovered that the wrong set of options were being used to validate
the SAML response.

npm test passes