test: add unit test for Swarm SAML validation
The new test is only the HAS side of the process, with the unit test making
the requests that Swarm would make. Tests the validation of the SAML
response sent from HAS to Swarm, then to p4d, and finally back to HAS again.

Added saml2-js dev dependency.

npm test passes