fix: quietly ignore missing OIDC secrets file
The configure script was failing if the OIDC secrets file named in the .env
file was missing. The example.env file that is copied into place by the
package post-install had a OIDC_CLIENT_SECRET_FILE setting that referred to
a file that does not exist, thus guaranteeing that the configure script
would fail immediately.

Also, change the example.env such that it no longer introduces any problems
like this. The only settings not commented out are those referring to the
TLS certificate files.