Add setting to assume client is authorized
In some scenarios, it is necessary for HAS to listen on HTTPS, but is also
very difficult to get the client certificate from the extension to pass
through reverse proxies. To alleviate this difficulty, this change adds a
new setting named ASSUME_CLIENT_AUTHORIZED which does exactly what the name
implies: assume that the client asking for the login request status is
authorized to make that query.

doc: undoc

npm test passes