fix: ensure defined SAML settings override IdP metadata
The whole point of defining SAML settings in the configuration is for them
to take precedence over anything else. This was working at some point in the
past, but has since broken. For instance, the IdP advertises several certs
for signing SAML assertions, but the node-saml library is unable to process
this for whatever reason. As a work-around, the admin can set the
IDP_CERT_FILE setting to a single certificate value.

npm test passes