Add logging useful for configuring TRUST_PROXY
Configuring HAS to trust a reverse proxy can be tricky, but with the logging
about that DEBUG_PROXY enables it should be easier. Simply set DEBUG_PROXY
to any value in the HAS configuration, restart, and perform a SAML login.
Then look in the logs for "saml: login request" to see the HTTP request
information related to the reverse proxy. Of particular interest are the
headers whose names start with "X-Forwarded-", as well as the remote address.

npm test passes