feat: add basic user/group provisioning suport
Imported from a separate repository, this change introduces a new feature to
HAS with support for provisioning users and groups in Helix Core Server, via
the SCIM protocol. Minimal testing with Azure AD, Okta, and OneLogin. This
new feature is not yet tested or supported, and lacks documentation and
scripted configuration.

With respect to support for SCIM, this service does not support renaming
groups (Helix Core lacks a `p4 renamegroup`), nor does it do anything when a
user is marked as inactive (Helix Core does not have an "active" state for
users). Otherwise, basic user and group management is functional.

The new configuration settings for the SCIM endpoint are BEARER_TOKEN,
P4PORT, P4USER, and P4PASSWD, all of which are described briefly in the
example.env file.

npm test passes