Switch to ECMAScript Modules from CommonJS
JavaScript (as defined by the ECMAScript standard) has had a new standard
for module support for some time now, and Node.js has supported this since
late 2020. At least one of the third party modules has now switched to this
new module system. As a result, using the latest version of that module is
no longer feasible without updating the entire application. We can expect
more libraries to do the same, so now is the time to make this change.

What this change entails:

* All require('foo') calls are now "import 'foo'"
* Dropped the 'include()' hack in favor of import 'helix-auth-svc/lib/...'
* Format of logging.config.js and saml_idp.conf.js changed to ESM format
  (now 'export default = {}' instead of 'module.exports = {}')
* bin/www had to be renamed to bin/www.js, but should not matter much as
  Node.js will add the '.js' to 'bin/www' automatically

npm test passes