Package builds no longer use the single binary
There are several issues with using the pkg utility to produce a single
binary of the application, its dependencies, and the Node.js binary. In
short, it does not fully support the ECMAScript 6 standard, which has been
out since 2015. The only drawback of this change is that the package now
requires that the `nodejs` package be installed beforehand. This had been
the case several releases back, so in a way this is a regression.

One advantage is that debugging a HAS installation by adding logging to the
source code, or providing a hot fix to a new issue, is possible once again.