Use CommonJS format for configuration files
The logging and Swarm/SAML configuration files should continue to use the
CommonJS format, avoiding the need for the administrator to convert their
existing files to the new format. However, the names of the files must be
changed to have a .cjs extension. Will attempt to address the name change
with the configuration script and package post-install script.