Strip comments and blank lines when creating .env
When creating the initial .env file, either in the install.sh script or in
the package post-install, use the example.env file but without the comments
and blank lines. Leaving the comments is especially confusing because the
commented out settings are present, while the new setting is at the bottom
of the file. This situation makes it very easy for everyone to miss the fact
that the settings are defined. By stripping away all of the fluff, the
configuration file becomes short and easy to read.